+++ arrow-mtl-data/src/test/kotlin/arrow/mtl/test/CanTest.kt
@@ -1,4 +1,4 @@
-package arrow.smash
+package arrow.mtl.test
 
 import arrow.core.Invalid
 import arrow.core.Ior
@@ -6,24 +6,14 @@ import arrow.core.None
 import arrow.core.Option
 import arrow.core.Some
 import arrow.core.Valid
-import arrow.smash.extensions.can.applicative.applicative
-import arrow.smash.extensions.can.bicrosswalk.bicrosswalk
-import arrow.smash.extensions.can.bifunctor.bifunctor
-import arrow.smash.extensions.can.bitraverse.bitraverse
-import arrow.smash.extensions.can.eq.eq
-import arrow.smash.extensions.can.eqK.eqK
-import arrow.smash.extensions.can.eqK2.eqK2
-import arrow.smash.extensions.can.functor.functor
-import arrow.smash.extensions.can.hash.hash
-import arrow.smash.extensions.can.monad.monad
-import arrow.smash.extensions.can.show.show
-import arrow.smash.extensions.can.traverse.traverse
 import arrow.core.extensions.eq
 import arrow.core.extensions.hash
 import arrow.core.extensions.semigroup
 import arrow.core.extensions.show
+import arrow.mtl.test.generators.can
+import arrow.core.left
+import arrow.core.right
 import arrow.core.test.UnitSpec
-import arrow.smash.generators.can
 import arrow.core.test.laws.BicrosswalkLaws
 import arrow.core.test.laws.BifunctorLaws
 import arrow.core.test.laws.BitraverseLaws
@@ -33,8 +23,35 @@ import arrow.core.test.laws.MonadLaws
 import arrow.core.test.laws.ShowLaws
 import arrow.core.test.laws.TraverseLaws
 import arrow.core.toT
-import arrow.smash.generators.genK
-import arrow.smash.generators.genK2
+import arrow.mtl.Can
+import arrow.mtl.CanOf
+import arrow.mtl.CanPartialOf
+import arrow.mtl.component1
+import arrow.mtl.component2
+import arrow.mtl.extensions.can.applicative.applicative
+import arrow.mtl.extensions.can.bicrosswalk.bicrosswalk
+import arrow.mtl.extensions.can.bifunctor.bifunctor
+import arrow.mtl.extensions.can.bitraverse.bitraverse
+import arrow.mtl.extensions.can.eq.eq
+import arrow.mtl.extensions.can.eqK.eqK
+import arrow.mtl.extensions.can.eqK2.eqK2
+import arrow.mtl.extensions.can.functor.functor
+import arrow.mtl.extensions.can.hash.hash
+import arrow.mtl.extensions.can.monad.monad
+import arrow.mtl.extensions.can.show.show
+import arrow.mtl.extensions.can.traverse.traverse
+import arrow.mtl.getLeftOrElse
+import arrow.mtl.getOrElse
+import arrow.mtl.leftOption
+import arrow.mtl.rightOption
+import arrow.mtl.test.generators.genK
+import arrow.mtl.test.generators.genK2
+import arrow.mtl.toCan
+import arrow.mtl.toIor
+import arrow.mtl.toLeftCan
+import arrow.mtl.toRightCan
+import arrow.mtl.toValidated
+import arrow.mtl.toValidatedLeft
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Monad
 import io.kotlintest.properties.Gen
@@ -237,21 +254,21 @@ class CanTest : UnitSpec() {
       )
     }
 
-    "left() returns Option<A>" {
+    "leftOption() returns Option<A>" {
       allEqual<Int, String>(
-        { neither.left() to None },
-        { left.left() to Some(a) },
-        { right.left() to None },
-        { both.left() to Some(a) }
+        { neither.leftOption() to None },
+        { left.leftOption() to Some(a) },
+        { right.leftOption() to None },
+        { both.leftOption() to Some(a) }
       )
     }
 
-    "right() returns Option<B>" {
+    "rightOption() returns Option<B>" {
       allEqual<Int, String>(
-        { neither.right() to None },
-        { left.right() to None },
-        { right.right() to Some(b) },
-        { both.right() to Some(b) }
+        { neither.rightOption() to None },
+        { left.rightOption() to None },
+        { right.rightOption() to Some(b) },
+        { both.rightOption() to Some(b) }
       )
     }
