+++ arrow-mtl-data/src/test/kotlin/arrow/mtl/test/generators/GenK2.kt
@@ -1,13 +1,14 @@
-package arrow.smash.generators
+package arrow.mtl.test.generators
 
 import arrow.Kind2
 import arrow.core.test.generators.GenK2
-import arrow.smash.Can
-import arrow.smash.ForCan
+import arrow.mtl.Can
+import arrow.mtl.ForCan
 import io.kotlintest.properties.Gen
 
 fun Can.Companion.genK2() =
   object : GenK2<ForCan> {
+    @Suppress("UNCHECKED_CAST")
     override fun <A, B> genK(genA: Gen<A>, genB: Gen<B>): Gen<Kind2<ForCan, A, B>> =
       Gen.can(genA, genB) as Gen<Kind2<ForCan, A, B>>
   }
