--- arrow-smash/build.gradle
@@ -1,27 +1,0 @@
-plugins {
-    id "maven-publish"
-    id "base"
-    id "org.jetbrains.kotlin.jvm"
-    id "org.jetbrains.kotlin.kapt"
-    id "net.rdrei.android.buildtimetracker"
-    id "org.jetbrains.dokka"
-    id "org.jlleitschuh.gradle.ktlint"
-    id "ru.vyarus.animalsniffer"
-}
-
-apply from: "$SUBPROJECT_CONF"
-apply from: "$DOC_CONF"
-apply from: "$PUBLISH_CONF"
-
-dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$KOTLIN_VERSION"
-    compile "io.arrow-kt:arrow-core-data:$VERSION_NAME"
-    compile "io.arrow-kt:arrow-annotations:$VERSION_NAME"
-    kapt "io.arrow-kt:arrow-meta:$VERSION_NAME"
-    kaptTest "io.arrow-kt:arrow-meta:$VERSION_NAME"
-    compileOnly "io.arrow-kt:arrow-meta:$VERSION_NAME"
-    testCompileOnly "io.arrow-kt:arrow-meta:$VERSION_NAME"
-    testRuntime "org.junit.vintage:junit-vintage-engine:$JUNIT_VINTAGE_VERSION"
-    testCompile "io.kotlintest:kotlintest-runner-junit5:$KOTLIN_TEST_VERSION", excludeArrow
-    testCompile project(":arrow-incubator-test")
-}
