+++ settings.gradle
@@ -13,4 +13,3 @@ include 'arrow-kindedj'
 include 'arrow-validation'
 include 'arrow-fx-mtl'
 include 'arrow-incubator-test'
-include 'arrow-smash'
